Fix concurrency limitations in process_queue/1.
We forgot to handle the concurrency limit when we dequeue from the
queue. The test case found this in a poll, because when you ask the
queue to poll, a token gets added. Normally, the concurrency limit
should make this token stay in the bucket, but what happened was that
the token got eaten by the item in the queue since the conc. limit was
not tested.