Fix another Model mistake.
The problem here was that we used Rate rather than BucketCount when we
replenish tokens. It makes sense that if we configure the bucket with
a rate above the token limit, then the rate is effectively limited by
the token limit parameter. The SUT acts like this, and now the model
reflects this kind of setup.