Push down a timestamp into the safetyvalve code.
The model now pushes a timestamp down into the code and the code reads that timestamp.
But it doesn't use it right now. It just throws it away. But it does makes sure that we can
handle timestamps inside the code and we know have something to use it for.

Going forward, we can add an expiration step to the queue when replenishing stuff.