When enqueuing, increase the time_point in the model.
This change is needed because it reflects reality. The rule is that each enqueue operation
must be uniquely generated, e.g., from an erlang:now() timestamp or a pairing
{os:timestamp(), self()} which a unique total ordering.

I think this is an acceptable decision. In normal operation, this should not be a problem and
if it proves to be, there is a solution to the problem, simply by enforcing uniqueness.