Introduce code to run the CoDel queue base.
This eqc test will just run the queue aggressively in order to test that
it really works like it was intended for it to work. This makes sure there
are no ways by which we can force the queue to crash.

In other words, we just proved a termination property.