Improve the Chaos Monkey™ kill jobs in the asking state.
This branch is an attempt at identifying Issue #1. We arrange it such that we
can kill jobs which are currently 'asking' for the ability to run. This leads to
severe inconsistency between the model and the underlying implementation.

The right solution is really a question of semantics, and we don't a priori know
what the right course of action happens to be.