Adapt the code base to R18+ Time API.
Timestamps now force the use of a unique integer. While far from
perfect, it obeys the old rules of the queues and thus it works the
same way as before. The speed is exactly as slow as before, by relying
on something like erlang:now().

While here, go through all the CoDel code and fix up the code base.